RasterSource.gridExtent[Long] instead of rasterExtent

+++ vlm/src/test/scala/geotrellis/contrib/vlm/spark/RasterSourceRDDSpec.scala
@@ -193,7 +193,7 @@ class RasterSourceRDDSpec extends FunSpec with TestEnvironment with BetterRaster
 
     val multibandTilePath = s"${new File("").getAbsolutePath()}/src/test/resources/img/aspect-tiled-0-1-2.tif"
 
-    val noDataTile = ArrayTile.alloc(cellType, rasterSource.cols, rasterSource.rows).fill(NODATA).interpretAs(cellType)
+    val noDataTile = ArrayTile.alloc(cellType, rasterSource.cols.toInt, rasterSource.rows.toInt).fill(NODATA).interpretAs(cellType)
 
     val paths: Seq[String] =
       0 to 5 map { index =>
